feat: client 화면 및 공통 컴포넌트 기준으로 폴더 재분류 후 관련 styles, hooks 를 폴더 내 정리
