+++ src/Components/index.tsx
@@ -1,0 +1,29 @@
+import Loading from "./Loading";
+import LoadingBall from "./LoadingBall";
+import PageHeader from "./PageHeader";
+import Chart from "./Chart";
+import Nav from "./Nav/Nav";
+import Modal from "./Modals/Modal";
+import Images from "./Images/Images";
+import Noimage from "./Images/Noimage";
+import ToggleButton from "./Buttons/ToggleButton";
+import ButtonDefaultStyle from "./Buttons/ButtonDefault";
+import InputDefault from "./Form/InputDefault";
+import LabelDefault from "./Form/LabelDefault";
+import TextareaDefault from "./Form/TextareaDefault";
+
+export {
+  Loading,
+  LoadingBall,
+  PageHeader,
+  Chart,
+  Nav,
+  Modal,
+  Images,
+  Noimage,
+  ToggleButton,
+  ButtonDefaultStyle,
+  InputDefault,
+  LabelDefault,
+  TextareaDefault
+};
