+++ src/Page/Client/ClientSelectList/index.tsx
@@ -1,99 +1,24 @@
 import { useState } from "react";
 import { useRecoilState } from "recoil";
-import styled from "styled-components";
+import { useNavigate, useParams } from "react-router-dom";
 import { IoIosAddCircle } from "react-icons/io";
 import { AiFillMinusCircle } from "react-icons/ai";
-import { selectMenuListState } from "../../lib/state/productItemState";
-import { IOrderSelectedItem } from "./ClientMenu";
-import Modal from "../../Components/Modals/Modal";
-import { Headline1, SubTitle1, SubTitle2, Body1 } from "../../lib/styles/mixin";
-import OrderStateBar from "./OrderStateBar";
-import PaymentModalChildren from "./Modal/PaymentModalChildren";
-import { useNavigate, useParams } from "react-router-dom";
-import ButtonDefaultStyle from "../../Components/Buttons/ButtonDefault";
-import { translateLocalCurrency } from "../../lib/utils/helper/translateLocalCurrency";
-import Noimage from "../../Components/Images/Noimage";
-import Images from "../../Components/Images/Images";
-
-const Header = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  h1 {
-    ${Headline1}
-    line-height: 1;
-    ${({ theme }) => theme.device.mobile} {
-      font-size: 4.2rem;
-    }
-  }
-  button {
-    cursor: pointer;
-    padding: 0.5rem 2rem;
-    border: 0;
-    font-size: 2.8rem;
-    color: ${(props) => props.theme.color.fontColorWhite};
-    border-radius: 0.3rem;
-    background-color: ${(props) => props.theme.color.gray300};
-  }
-`;
-const MenuListWrapper = styled.div`
-  margin: 1rem 0;
-`;
-const MenuListItem = styled.div`
-  display: grid;
-  grid-template-columns: 20% 50% 30%;
-  padding: 0.5rem 0;
-  border-bottom: 1px solid black;
-  img {
-    border: 1px solid;
-    border-color: ${(props) => props.theme.color.gray300};
-  }
-  div {
-    /* display: grid; */
-    padding: 0 0.5rem;
-  }
-  h2 {
-    ${SubTitle1}
-    font-weight: bold;
-    padding: 1.3rem 0 0.8rem 0;
-  }
-  p {
-    ${Body1}
-    padding: 0.3rem 0;
-  }
-  span {
-    vertical-align: super;
-    padding: 0.5rem;
-  }
-  .price {
-    margin-top: 1rem;
-    ${SubTitle2}
-  }
-`;
-const DeleteButton = styled(ButtonDefaultStyle)`
-  background-color: ${(props) => props.theme.color.gray300};
-`;
-const ResetButtonWrapper = styled.div`
-  display: flex;
-  justify-content: flex-end;
-`;
-const ResetButton = styled(ButtonDefaultStyle)`
-  text-align: center;
-  background-color: ${(props) => props.theme.color.primary500};
-`;
-export const AddCountButton = styled(ButtonDefaultStyle)`
-  background-color: unset;
-  color: ${(props) => props.theme.color.primary500};
-  font-size: 2.3rem;
-  vertical-align: middle;
-`;
-export const MinusCountButton = styled(ButtonDefaultStyle)`
-  background-color: unset;
-  color: ${(props) => props.theme.color.error500};
-  font-size: 2.3rem;
-  vertical-align: middle;
-`;
+import { Images, Modal, Noimage } from "../../../Components";
+import PaymentModalChildren from "../Modal/PaymentModalChildren";
+import { selectMenuListState } from "../../../lib/state/productItemState";
+import { IOrderSelectedItem } from "../ClientMenu";
+import {
+  Header,
+  MenuListWrapper,
+  MenuListItem,
+  DeleteButton,
+  ResetButtonWrapper,
+  ResetButton,
+  AddCountButton,
+  MinusCountButton
+} from "./styles";
+import { translateLocalCurrency } from "../../../lib/utils/helper/translateLocalCurrency";
+import OrderStateBar from "../OrderStateBar";
 
 const ClientSelectList = () => {
   const navigate = useNavigate();
